Variance and averages calculations are now performed on perf-results

+++ src/Vector3D.cpp
@@ -178,7 +178,7 @@ Vector3D<S> Vector3D<S>::cross(Vector3D<S> a){
 template<class S>
 S Vector3D<S>::mag(){
    #ifndef TINSEL
-   return sqrt(_x*_x + _y*_y + _z*_z);
+   return sqrtf(_x*_x + _y*_y + _z*_z);
    #else
    return newt_sqrt(_x*_x + _y*_y + _z*_z);
    #endif
